+++ src/plskern.jl
@@ -171,8 +171,7 @@ end
 """ 
     transf(object::Union{Plsr, Splsr}, 
         X; nlv = nothing)
-Compute latent variables (LVs = scores T) from 
-    a fitted model.
+Compute latent variables (LVs = scores T) from a fitted model.
 * `object` : The fitted model.
 * `X` : Matrix (m, p) for which LVs are computed.
 * `nlv` : Nb. LVs to consider.